+++ scripts/lib/utilities/itemUtils.js
@@ -3,6 +3,7 @@ import {actorUtils, effectUtils, genericUtils, socketUtils, errors, compendiumUt
 import {gambitPremades} from '../../integrations/gambitsPremades.js';
 import {miscPremades} from '../../integrations/miscPremades.js';
 import {custom} from '../../events/custom.js';
+import {ItemMedkit} from '../../applications/medkit-item.js';
 function getSaveDC(item) {
     if (item.hasSave) return item.getSaveDC();
     let spellDC;
@@ -166,6 +167,9 @@ function isWeaponProficient(item) {
     if (item.actor.system.traits.weaponProf.value.has(CONFIG.DND5E.weaponProficienciesMap[item.system.type.value])) return true;
     return false;
 }
+async function itemUpdate(item) {
+    return await ItemMedkit.itemUpdate(item);
+}
 export let itemUtils = {
     getSaveDC,
     createItems,
@@ -187,5 +191,6 @@ export let itemUtils = {
     getSavedCastData,
     getGenericFeatureConfig,
     getItemByGenericFeature,
-    isWeaponProficient
+    isWeaponProficient,
+    itemUpdate
 };